+++ examples/stand_alone/diagnostics/diagnostics.cc
@@ -1,0 +1,62 @@
+/*
+ * Copyright (C) 2017 Open Source Robotics Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+*/
+
+#include <gazebo/msgs/msgs.hh>
+#include <gazebo/transport/transport.hh>
+#include <gazebo/util/Diagnostics.hh>
+#include <gazebo/gazebo_client.hh>
+
+#include <iostream>
+
+/////////////////////////////////////////////////
+// Function is called everytime a message is received.
+void cb(ConstWorldStatisticsPtr &_msg)
+{
+  DIAG_TIMER_STOP("example_world_stats")
+  // Dump the message contents to stdout.
+  std::cout << _msg->DebugString();
+  DIAG_TIMER_START("example_world_stats")
+}
+
+/////////////////////////////////////////////////
+int main(int _argc, char **_argv)
+{
+  // Load gazebo
+  gazebo::client::setup(_argc, _argv);
+
+  // Create our node for communication
+  gazebo::transport::NodePtr node(new gazebo::transport::Node());
+  node->Init();
+
+  // Start the diagnostic timer
+  DIAG_TIMER_START("example_world_stats")
+
+  // Listen to Gazebo world_stats topic
+  gazebo::transport::SubscriberPtr sub = node->Subscribe("~/world_stats", cb);
+
+  // Wait for 3 seconds
+  for (int i = 0; i < 3000; ++i)
+  {
+    gazebo::common::Time::MSleep(1);
+  }
+
+  // Start the diagnostic timer
+  DIAG_TIMER_STOP("example_world_stats")
+
+  // Make sure to shut everything down.
+  gazebo::client::shutdown();
+}