+++ Changelog.md
@@ -291,6 +291,9 @@
 
 ## Gazebo 7.x.x (2017-xx-xx)
 
+1. Diagnostics: enable test and don't create so many empty folders
+    * [Pull request 2798](https://bitbucket.org/osrf/gazebo/pull-requests/2798)
+
 1. Parallelize ODE physics with threaded islands parameter
     * [Pull request 2775](https://bitbucket.org/osrf/gazebo/pull-requests/2775)